docs(utils): add documentation for string formatting logic (#517)
