+++ television/utils/strings.rs
@@ -1,3 +1,5 @@
+use lazy_regex::{regex, Lazy, Regex};
+
 /// Returns the index of the next character boundary in the given string.
 ///
 /// If the given index is already a character boundary, it is returned as is.
@@ -560,6 +562,39 @@ pub fn shrink_with_ellipsis(s: &str, max_length: usize) -> String {
     let second_half =
         slice_at_char_boundaries(s, s.len() - half_max_length, s.len());
     format!("{first_half}…{second_half}")
+}
+
+pub static CMD_RE: &Lazy<Regex> = regex!(r"\{(\d+)\}");
+
+/// Formats a prototype string with the given template and source strings.
+///
+/// # Example
+/// ```
+/// use television::utils::strings::format_string;
+///
+/// let template = "cat {} {1}";
+/// let source = "foo:bar:baz";
+/// let delimiter = ":";
+///
+/// let formatted = format_string(template, source, delimiter);
+/// assert_eq!(formatted, "cat 'foo:bar:baz' 'bar'");
+/// ```
+pub fn format_string(template: &str, source: &str, delimiter: &str) -> String {
+    let parts = source.split(delimiter).collect::<Vec<&str>>();
+
+    let mut formatted_string =
+        template.replace("{}", format!("'{}'", source).as_str());
+
+    formatted_string = CMD_RE
+        .replace_all(&formatted_string, |caps: &regex::Captures| {
+            let index =
+                // these unwraps are safe because of the regex pattern
+                caps.get(1).unwrap().as_str().parse::<usize>().unwrap();
+            format!("'{}'", parts.get(index).unwrap_or(&""))
+        })
+        .to_string();
+
+    formatted_string
 }
 
 #[cfg(test)]
